Move ThanksPage to containers
