+++ src/routes.js
@@ -7,7 +7,7 @@ import { loadSession, trackPage } from './actions/sessionActions'
 import { loadOrganization } from './actions/navActions.js'
 import Home from './containers/home'
 import SignPetition from './containers/sign-petition'
-import ThanksPage from './pages/thanks'
+import ThanksPage from './containers/thanks'
 import SearchPage from './pages/search'
 import PetitionCreatorDashboard from './containers/petition-creator-dashboard'
 import CreatePetitionPage from './pages/create-petition-page'
